+++ components/Card/ProjectCard.tsx
@@ -1,3 +1,4 @@
+import { Button, Icon, Scrim, Tooltip } from "@equinor/eds-core-react";
 import React, { useState } from "react";
 import { faveProject, unfaveProject } from "services/projectApi";
 import { useAccount, useMsal } from "@azure/msal-react";
@@ -5,26 +6,30 @@ import { useMutation, useQueryClient } from "react-query";
 
 import { AccessBox } from "components/AccessBox";
 import Heart from "components/Heart";
-import Labels from "components/Labels";
+import Labels from "components/Labels/Labels";
+import ManageLabelBox from "components/Labels/ManageLabelBox";
 import ProjectCardHeader from "./ProjectCardHeader";
-import { Scrim } from "@equinor/eds-core-react";
 import { UserDots } from "./UserDots";
 import { getMyAccess } from "utils/getMyAccess";
-import { getOwner } from "utils/getOwner";
 import styles from "./Card.module.scss";
+import { tag } from "@equinor/eds-icons";
 import { useRouter } from "next/router";
 import { vsmProject } from "../../interfaces/VsmProject";
 
 export function ProjectCard(props: { vsm: vsmProject }): JSX.Element {
+  const { userIdentity: createdBy } = props.vsm.created;
   const queryClient = useQueryClient();
   const [isMutatingFavourite, setIsMutatingFavourite] = useState(false);
   const router = useRouter();
 
   const [visibleScrim, setVisibleScrim] = useState(false);
+  const [visibleLabelScrim, setVisibleLabelScrim] = useState(false);
+
   const { accounts } = useMsal();
   const account = useAccount(accounts[0] || {});
   const myAccess = getMyAccess(props.vsm, account);
-  const isAdmin = myAccess === "Admin" || myAccess === "Owner";
+  const isAdmin = myAccess === "Admin";
+  const userCanEdit = isAdmin || myAccess == "Contributor";
 
   const handleSettled = () => {
     queryClient.invalidateQueries().then(() => setIsMutatingFavourite(false));
@@ -65,7 +70,7 @@ export function ProjectCard(props: { vsm: vsmProject }): JSX.Element {
         onClick={() => router.push(`/process/${props.vsm.vsmProjectID}`)}
       >
         <div className={styles.card}>
-          <div className={styles.topSection}>
+          <div className={styles.section}>
             <ProjectCardHeader vsm={props.vsm} />
             <Heart
               isFavourite={props.vsm.isFavorite}
@@ -74,25 +79,55 @@ export function ProjectCard(props: { vsm: vsmProject }): JSX.Element {
               isLoading={isMutatingFavourite}
             />
           </div>
-          <div className={styles.bottomSection}>
+          <div className={`${styles.section} ${styles.labelSection}`}>
             <Labels labels={props.vsm.labels} />
-            <UserDots
-              users={[...props.vsm.userAccesses.map((u) => u.user)]}
-              setVisibleScrim={(any: boolean) => setVisibleScrim(any)}
-            />
+          </div>
+          <div className={`${styles.section} ${styles.bottomSection}`}>
+            {createdBy && (
+              <UserDots
+                users={[
+                  `${createdBy}`,
+                  ...props.vsm.userAccesses.map((u) => u.user),
+                ]}
+                setVisibleScrim={(any: boolean) => setVisibleScrim(any)}
+              />
+            )}
+            {userCanEdit && (
+              <Tooltip title="Manage process labels" placement="right">
+                <Button
+                  color="primary"
+                  variant="ghost_icon"
+                  style={{ height: "30px", width: "30px" }}
+                  onClick={(e) => {
+                    e.stopPropagation();
+                    setVisibleLabelScrim(true);
+                  }}
+                >
+                  <Icon data={tag} />
+                </Button>
+              </Tooltip>
+            )}
           </div>
         </div>
       </button>
 
-      {visibleScrim && (
-        <Scrim onClose={() => setVisibleScrim(false)} isDismissable>
-          <AccessBox
-            project={props.vsm}
-            handleClose={() => setVisibleScrim(false)}
-            isAdmin={isAdmin}
-          />
-        </Scrim>
-      )}
+      <Scrim
+        open={visibleScrim}
+        onClose={() => setVisibleScrim(false)}
+        isDismissable
+      >
+        <AccessBox
+          project={props.vsm}
+          handleClose={() => setVisibleScrim(false)}
+          isAdmin={isAdmin}
+        />
+      </Scrim>
+
+      <ManageLabelBox
+        handleClose={() => setVisibleLabelScrim(false)}
+        isVisible={visibleLabelScrim}
+        process={props.vsm}
+      />
     </>
   );
 }