+++ components/ErrorDialog.tsx
@@ -1,7 +1,7 @@
+import { Button, Scrim } from "@equinor/eds-core-react";
+
 import React from "react";
 import { unknownErrorToString } from "utils/isError";
-import { Button, Scrim } from "@equinor/eds-core-react";
-import { restore } from "@equinor/eds-icons";
 
 export function ErrorDialog({
   visible,
@@ -16,7 +16,7 @@ export function ErrorDialog({
 
   return (
     error && (
-      <Scrim isDismissable onClose={onClose}>
+      <Scrim open isDismissable onClose={onClose}>
         <div
           style={{
             backgroundColor: "white",