+++ interfaces/VsmProject.tsx
@@ -4,13 +4,16 @@ import { vsmObject } from "./VsmObject";
 import { vsmObjectTypes } from "../types/vsmObjectTypes";
 import { vsmTaskTypes } from "../types/vsmTaskTypes";
 
+// Todo: Structure is out of date. Update to match new structure.
 export interface vsmProject {
   vsmProjectID: number;
   name: string;
   toBeProcessID?: number;
   currentProcessId?: number;
   labels: processLabel[];
-  created: string; // date
+  created: {
+    userIdentity: string; // date
+  };
   lastUpdated: string; // date
   objects: Array<vsmObject>;
   userAccesses: Array<userAccess>;