+++ interfaces/queryObject.tsx
@@ -1,0 +1,5 @@
+export interface queryObject {
+  orderBy?: undefined | string;
+  q?: undefined | string; // search query (as defined in Swagger)
+  rl?: undefined | string | string[]; // label ID array (as defined in Swagger)
+}